[REFACTOR] app.js 리팩토링
  app.js와 bin.js의 역할을 좀더 명확하게 구분
